+++ Leek/core/src/main/java/club/cybercraftman/leek/core/strategy/common/Signal.java
@@ -39,7 +39,7 @@ public class Signal {
     /**
      * 量
      */
-    private Integer volume;
+    private Long volume;
 
     /**
      * 乘数
